+++ lib/mocks.dart
@@ -1,0 +1,49 @@
+import 'package:geek_plants/pathStrings.dart';
+
+import 'model/plant.dart';
+
+List categories = [
+  "Все",
+  "Комнатные",
+  "Уличные",
+  "Цветущие",
+  "Однолетние",
+  "Многолетние"
+];
+List plantList = [
+  Plant(
+    name: "Абутилон (Катаника)",
+    latName: "лат.Abutilon",
+    photoPath: abutilonPath,
+  ),
+  Plant(
+    name: "Агава",
+    latName: "лат.Agave",
+    photoPath: abutilonPath,
+  ),
+  Plant(
+    name: "Бальзамин",
+    latName: "лат.Impatience",
+    photoPath: abutilonPath,
+  ),
+  Plant(
+    name: "Бальзамин",
+    latName: "лат.Impatience",
+    photoPath: abutilonPath,
+  ),
+  Plant(
+    name: "Бальзамин",
+    latName: "лат.Impatience",
+    photoPath: abutilonPath,
+  ),
+  Plant(
+    name: "Бальзамин",
+    latName: "лат.Impatience",
+    photoPath: abutilonPath,
+  ),
+  Plant(
+    name: "Бальзамин",
+    latName: "лат.Impatience",
+    photoPath: abutilonPath,
+  ),
+];
